requirements.txt
Flask
python-dotenv
auto-machine-learning
